remove bad files
