+++ JamNavigator/JamNavigator/PushNotificationHandler.swift
@@ -8,10 +8,28 @@
 import Foundation
 import Firebase
 import UserNotifications
+import UIKit
+import Amplify
 
-
-extension AppDelegate : UNUserNotificationCenterDelegate {
+extension UIViewController {
     
+    // リモート通知を要求する
+    func pushRemote() {
+        let message = #"{"message": "my new Todo"}"#
+        let request = RESTRequest(path: "/push/123", body: message.data(using: .utf8))
+        Amplify.API.post(request: request) {
+            result in
+            switch result {
+            case .success(let data):
+                let str = String(decoding: data, as: UTF8.self)
+                print("＋＋＋＋＋", str)
+                    
+            case .failure(let apiError):
+                print("＋＋＋＋＋", apiError)
+            }
+        }
+    }
+
     // ローカル通知を出す
     func pushLocal(title: String, body: String, delaySeconds: Double = 0.0, sound: UNNotificationSound = .default) {
         let content = UNMutableNotificationContent()
@@ -24,6 +42,10 @@ extension AppDelegate : UNUserNotificationCenterDelegate {
         UNUserNotificationCenter.current().add(request, withCompletionHandler: nil) // 通知登録
     }
 
+}
+
+extension AppDelegate : UNUserNotificationCenterDelegate {
+    
     // HACK:アプリがフォアグランドにいるときにPUSHされたイベント処理
     func userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
